console/requirements/base.txt
# Password storage
argon2-cffi==19.2.0

# Django
Django==2.2.13

# Forms
django-crispy-forms==1.8.1

# Django debug toolbar
django-debug-toolbar==2.2

# Django environ
django-environ==0.4.5

# Django Extensions
django-extensions==2.2.5

# Django filters for API
django-filter==2.2.0

# Django recurrences for periodic tasks
django-recurrence==1.10.1

# SAML
django_saml2_auth==2.2.1

# Django REST Framework
djangorestframework==3.9.2

# Yet another Swagger generator
drf-yasg==1.16.1

# fqdn
fqdn==1.4.0

# ipython
ipython

# packaging required for drf-yasg, adding this to be safe
packaging

# Python-PostgreSQL Database Adapter
psycopg2-binary==2.8.4

# Used for django-extensions graph_models
pygraphviz==1.5

# nmap
python-libnmap==0.7.0

# redis
redis==3.4.1

# python-rq - Redis Queue
rq==1.2.2

# XML to JSON
xmljson==0.2.0
